Read symbol name from provided registry
So far we have used the stock symbol as the name of each Ledger journal
entry. That's ultimately not what we want.
With this change we assume a user-provided registry -- a simple JSON
file containing a mapping from symbol to name -- that we use for looking
up a name corresponding to a symbol and use that instead.