Introduce -v/--verbose option to control tracing log level
In order to debug problems arising the alpaca crate which we use for
communicating with the Alpaca API supports tracing. However, tracing
needs to be explicitly initialized by the containing application in
order to do anything useful.
This change takes care of this initialization. The tracing log level can
be controlled through the newly introduced -v/--verbose option.