Switch to using structopt for parsing command line options
Our handling of command line arguments has been very ad-hoc so far. That
was okay because we only supported (and required) a single argument.
However, in the future we at least want to allow users to control the
tracing infrastructure and configure the log level.
To accommodate those cases, this change switches to a more sustainable
approach of handling arguments using structopt.