Move registry lookup into print_trade function
With this change we move the symbol to name lookup in the registry into
the print_trade function. In the future we also want to handle non-trade
activities where we need to do a similar lookup but conditionally. In
order to keep the code as consistent as possible, it makes more sense to
have this lookup happen inside the corresponding print function.