Unconditionally request all account activities
So far we were only interested in producing Ledger entries for actual
trade activities. In the future we want to account for dividends and
other non-trade activities.
In order to not have to list all activity types that we are interested
in and continuously updating this list, with this change we
unconditionally request all types of account activities.