Print activity data in descending order
So far we have printed activity data in the order we received it through
the Alpaca API -- which means old data first and more recent data later.
That approach has the downside that for our intents, namely a journal
file, the order is exactly reversed to what it should be, because the
most recent data would appear at the top (emitted last).
With this patch we change this state of affair by printing the most
recent data first, which makes it appear at the bottom of the output.