Add support for paging through all account activities
So far we were only able to retrieve the last 100 account activities.
That was a limitation of the apca crate. With version 0.18.0 the crate
added support for paging through all account activities and so with this
patch we adjust the logic to make use of this feature. This means that
the program will now report all of the user's activities.