Add support for regulatory fees
This change introduces rudimentary support for reporting regulatory
fees, which have recently been included in the Alpaca reported activity.
In its current state, a good amount of manual post-processing is
required to form the (most likely desired) Ledger entry. That is mostly
due to the simplicity of the API provided.