Handle activities in batches spanning at least one full day
With this change we process account activities in batches of a day,
meaning that we retrieve as many activity entries until we have at least
a full day (or there are none left). This change is not useful in and of
itself, but it will allow us to potentially merge partial fills and do
some other pre-processing before emitting the final activity log. Such
changes will come in the future.