Merge partial fills for the same order at the same price on a given day
The account activity stream reported by Alpaca contains separate entries
for each partial fill of an order. For large trades that can quickly get
cumbersome to handle.
With this change we attempt to merge partials fills for one and the same
ordering happening on the same day, as long as the fill price is the
same. Depending on the type of order used, these conditions may or may
not be fulfilled frequently, potentially resulting in a good reduction
of emitted ledger entries.