Use regular expressions for classifying fees
In order to somehow associate regulatory fees as reported by Alpaca with
the trades they belong to we unfortunately need to parse their
description to extract details not contained in the structured part of
the activity. We will use regular expressions for that purpose.
With this change we prepare for that step by introducing the
corresponding regular expressions and using them in the fee
classification function introduced earlier.