Try to associate regulatory fees with the corresponding trades
The reporting of regulatory fees introduced a while back required a lot
of manual work to match up and merge the emitted entries with the
corresponding trade transactions.
What a human would do in this case can mostly be described in the form
of an algorithm and automated and with this change we attempt to do that
on a best-effort basis: we parse the description associated with the
regulatory fee to infer the number of shares traded or total USD
proceeds, respectively, find a trade matching this criteria, and
associate the fee with it. Not that unfortunately there is a lot of room
for false association here. Different trades can happen to be for the
same number of shares or USD proceeds by happenstance and there is no
information about an order or even symbol of the security involved to
disambiguate. However, depending on the number of trades issued on any
single day the risk is presumably very small and it is very likely that
such problems will be caught when attempting to balance transactions
out.