Introduce --force-separate-fees option
Given that the association of regulatory fees to trades is a best-effort
endeavor with the possibility of errors, it is somewhat important that
users can opt out of it in case things go awry.
This change introduces a new option, --force-separate-fees, that can be
used to opt out of this matching up.