Add support for acquisitions
This change adds support for dealing with acquisition type non-trade
account activities. In case of an acquisition all shares will be sold at
a given price. The price is not contained in the activity object in
structured form, and instead we have to resort back to parsing it from
the description.