Fix merging of partial fills to no longer assume a specific ordering
As it turns out Alpaca cannot be trusted to deliver account activities
in strictly chronological order. That is, the "final" fill for an order
may actually be reported before other partial fills. That actually
breaks out partial fill merging logic.
To fix this problem, this change reworks the merge logic for partial
fills: instead of merging them into a strictly subsequent activity, we
now scan the entire activity array and merge with the final fill. The
downside of this approach is that if there is no final fill (either
because it never occurred or because it is simply not contained in the
batch of activities that we are processing), we may not merge partial
fills at all. That is assumed to be a minor nuisance, though, as it
should be a rare event (that is reasonably easy to fix up manually).