Add rudimentary support for handling stock splits
This change introduces rudimentary support for handling of stock splits
to the program. We refer to it as "rudimentary", because we ultimately
could attempt to find matching ADD and REMOVE shares entries and combine
them, but currently we merely emit one Ledger entry for each. That is
probably fine, given how infrequent the associated activity occurs in
practice -- it's not a big effort to touch things up manually.
Along with this change we bump the required version of the apca crate to
0.25.1. We need this more recent patch version in order to have the
price member available on non-trade account activities.