Add support for handling ADR fees
This change adds support for dealing with ADR fees. Because they are not
exactly reported in a structured manner, we have to infer the kind of
fee by parsing the descriptions, as we already do for other fees. This
patch extends the logic to recognize ADR fees and makes sure to emit the
correct record for them.