Warn on unsupported non-trade activity
So far we silently ignored unsupported non-trade activities. The main
reason for that behavior is that reporting an actual error would
potentially cause more harm and just logging a warning would mess up the
output.
As a compromise, with this change we log all traces to stderr while
starting to log such unsupported activities.